+++ src/main/kotlin/com/miyagi/shashin/configuration/WebSecurityConfig.kt
@@ -100,11 +100,12 @@ class MultiSecurityConfig {
             "/fonts/**",
             "/images/**",
             "/media/**",
-            "/users/register"
+            "/users/register",
+
         )
 
         var publicApiList = arrayOf(
-            "/api/v1/thumbnails/**", "/api/v1/image/**", "/api/v1/video/**", "/api/v1/profile/**", "/api/v1/share/**", "/api/v1/tags"
+            "/api/v1/thumbnails/**", "/api/v1/image/**", "/api/v1/video/**", "/api/v1/profile/**", "/api/v1/share/**", "/api/v1/tags", "/api/v1/metadata/**"
         )
 
         var publicList = publicApiList + resourceList + arrayOf(
@@ -124,7 +125,8 @@ class MultiSecurityConfig {
             "/*/rss",
             "/*/atom",
             "/users/login",
-            "/users/logout"
+            "/users/logout",
+            "/metadata/**"
         )
 
         private val adminApiList = arrayOf(